+++ CHANGELOG.md
@@ -3,6 +3,7 @@ This is an overview of major changes. Refer to the git repository for a full log
 Version 0.1.21
 -------------
 - Fixed #198 - Opening both .inselect files and images by drag-drop is broken (@quicklizard99)
+- Added #192 - Cookie cutter templates (@quicklizard99)
 - Fixed #48 - Zooming should follow mouse position (@quicklizard99)
 
 Version 0.1.20